requirements.txt
langchain==1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
chromadb==0.5.4
python-dotenv==1.2.1
pypdf==6.4.0
fastapi==0.123.0
uvicorn==0.38.0
